rgw: more headers cleanups
