+++ src/rgw/rgw_data_sync.h
@@ -9,7 +9,6 @@
 #include "common/RWLock.h"
 #include "common/ceph_json.h"
 
-
 #include "rgw_coroutine.h"
 #include "rgw_http_client.h"
 #include "rgw_bucket.h"
@@ -267,6 +266,8 @@ struct RGWDataSyncEnv {
   string status_oid();
 };
 
+class RGWRados;
+
 class RGWRemoteDataLog : public RGWCoroutinesManager {
   const DoutPrefixProvider *dpp;
   RGWRados *store;
@@ -284,12 +285,7 @@ class RGWRemoteDataLog : public RGWCoroutinesManager {
 
 public:
   RGWRemoteDataLog(const DoutPrefixProvider *dpp, RGWRados *_store,
-                   RGWAsyncRadosProcessor *async_rados)
-    : RGWCoroutinesManager(_store->ctx(), _store->get_cr_registry()),
-      dpp(dpp), store(_store), async_rados(async_rados),
-      http_manager(store->ctx(), completion_mgr),
-      lock("RGWRemoteDataLog::lock"), data_sync_cr(NULL),
-      initialized(false) {}
+                   RGWAsyncRadosProcessor *async_rados);
   int init(const string& _source_zone, RGWRESTConn *_conn, RGWSyncErrorLogger *_error_logger,
            RGWSyncTraceManager *_sync_tracer, RGWSyncModuleInstanceRef& module);
   void finish();
@@ -308,7 +304,6 @@ public:
 
 class RGWDataSyncStatusManager : public DoutPrefixProvider {
   RGWRados *store;
-  rgw_rados_ref ref;
 
   string source_zone;
   RGWRESTConn *conn;
@@ -375,7 +370,7 @@ public:
   }
 
   // implements DoutPrefixProvider
-  CephContext *get_cct() const override { return store->ctx(); }
+  CephContext *get_cct() const override;
   unsigned get_subsys() const override;
   std::ostream& gen_prefix(std::ostream& out) const override;
 };
@@ -512,11 +507,7 @@ public:
   RGWRemoteBucketLog(const DoutPrefixProvider *_dpp, RGWRados *_store,
                      RGWBucketSyncStatusManager *_sm,
                      RGWAsyncRadosProcessor *_async_rados,
-                     RGWHTTPManager *_http_manager)
-    : RGWCoroutinesManager(_store->ctx(), _store->get_cr_registry()),
-      dpp(_dpp), store(_store), status_manager(_sm),
-      async_rados(_async_rados), http_manager(_http_manager)
-  {}
+                     RGWHTTPManager *_http_manager);
 
   int init(const string& _source_zone, RGWRESTConn *_conn,
            const rgw_bucket& bucket, int shard_id,
@@ -558,13 +549,7 @@ class RGWBucketSyncStatusManager : public DoutPrefixProvider {
 
 public:
   RGWBucketSyncStatusManager(RGWRados *_store, const string& _source_zone,
-                             const rgw_bucket& bucket) : store(_store),
-                                                                                     cr_mgr(_store->ctx(), _store->get_cr_registry()),
-                                                                                     http_manager(store->ctx(), cr_mgr.get_completion_mgr()),
-                                                                                     source_zone(_source_zone),
-                                                                                     conn(NULL), error_logger(NULL),
-                                                                                     bucket(bucket),
-                                                                                     num_shards(0) {}
+                             const rgw_bucket& bucket);
   ~RGWBucketSyncStatusManager();
 
   int init();
@@ -576,7 +561,7 @@ public:
   static string obj_status_oid(const string& source_zone, const rgw_obj& obj); /* can be used by sync modules */
 
   // implements DoutPrefixProvider
-  CephContext *get_cct() const override { return store->ctx(); }
+  CephContext *get_cct() const override;
   unsigned get_subsys() const override;
   std::ostream& gen_prefix(std::ostream& out) const override;
